Update Openstack preset creation
