+++ src/app/settings/admin/presets/dialog/steps/settings/provider/openstack/template.html
@@ -17,7 +17,9 @@ limitations under the License.
       fxLayout="column"
       fxLayoutGap="8px">
 
-  <km-openstack-credentials [formControl]="form.get(Controls.Credentials)"></km-openstack-credentials>
+  <km-openstack-credentials [formControl]="form.get(Controls.Credentials)"
+                            (onChange)="onCredentialsChange($event)"
+                            [mode]="Modes.PresetCreation"></km-openstack-credentials>
 
   <div>OpenStack Configuration</div>
   <mat-form-field fxFlex>
