+++ src/app/settings/admin/presets/dialog/steps/settings/style.scss
@@ -18,8 +18,8 @@
 .km-provider-logo {
   background-position: left;
   margin: 15px 30px 10px 0;
-  max-height: 20px;
-  min-height: 20px;
+  max-height: 40px;
+  max-width: 80px;
 }
 
 .optional-section {
