+++ src/app/shared/components/openstack-credentials/application/component.ts
@@ -12,50 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {
-  ChangeDetectionStrategy,
-  ChangeDetectorRef,
-  Component,
-  forwardRef,
-  Input,
-  OnDestroy,
-  OnInit,
-  ViewChild,
-} from '@angular/core';
+import {ChangeDetectionStrategy, Component, forwardRef, OnDestroy, OnInit} from '@angular/core';
 import {FormBuilder, NG_VALIDATORS, NG_VALUE_ACCESSOR, Validators} from '@angular/forms';
-import {OpenstackCredentialsTypeService} from '@app/wizard/step/provider-settings/provider/extended/openstack/service';
 import {ClusterSpecService} from '@core/services/cluster-spec';
-import {DatacenterService} from '@core/services/datacenter';
 import {PresetsService} from '@core/services/wizard/presets';
-import {FilteredComboboxComponent} from '@shared/components/combobox/component';
 import {CloudSpec, Cluster, ClusterSpec, OpenstackCloudSpec} from '@shared/entity/cluster';
-import {OpenstackFloatingIPPool} from '@shared/entity/provider/openstack';
 import {NodeProvider} from '@shared/model/NodeProviderConstants';
 import {BaseFormValidator} from '@shared/validators/base-form.validator';
-import _ from 'lodash';
-import {EMPTY, merge, Observable, onErrorResumeNext} from 'rxjs';
-import {
-  catchError,
-  debounceTime,
-  distinctUntilChanged,
-  filter,
-  map,
-  switchMap,
-  take,
-  takeUntil,
-  tap,
-} from 'rxjs/operators';
+import {merge} from 'rxjs';
+import {debounceTime, distinctUntilChanged, filter, takeUntil} from 'rxjs/operators';
 
 enum Controls {
   ApplicationCredentialID = 'applicationCredentialID',
   ApplicationCredentialSecret = 'applicationCredentialSecret',
-  FloatingIPPool = 'floatingIPPool',
-}
-
-enum FloatingIPPoolState {
-  Ready = 'Floating IP Pool',
-  Loading = 'Loading...',
-  Empty = 'No Floating IP Pools Available',
 }
 
 @Component({
@@ -77,24 +46,13 @@ enum FloatingIPPoolState {
 })
 export class OpenstackApplicationCredentialsComponent extends BaseFormValidator implements OnInit, OnDestroy {
   private readonly _debounceTime = 250;
-  private _isFloatingPoolIPEnforced = false;
-  @ViewChild('floatingIPPoolCombobox')
-  private readonly _floatingIPPoolCombobox: FilteredComboboxComponent;
-
-  @Input() verifyCredentials: boolean;
-
   readonly Controls = Controls;
   isPresetSelected = false;
-  floatingIPPools: OpenstackFloatingIPPool[] = [];
-  floatingIPPoolsLabel = FloatingIPPoolState.Empty;
 
   constructor(
     private readonly _builder: FormBuilder,
     private readonly _presets: PresetsService,
-    private readonly _clusterSpecService: ClusterSpecService,
-    private readonly _datacenterService: DatacenterService,
-    private readonly _credentialsTypeService: OpenstackCredentialsTypeService,
-    private readonly _cdr: ChangeDetectorRef
+    private readonly _clusterSpecService: ClusterSpecService
   ) {
     super('Openstack Provider Basic');
   }
@@ -103,7 +61,6 @@ export class OpenstackApplicationCredentialsComponent extends BaseFormValidator 
     this.form = this._builder.group({
       [Controls.ApplicationCredentialID]: this._builder.control('', Validators.required),
       [Controls.ApplicationCredentialSecret]: this._builder.control('', Validators.required),
-      [Controls.FloatingIPPool]: this._builder.control(''),
     });
 
     merge(
@@ -113,15 +70,6 @@ export class OpenstackApplicationCredentialsComponent extends BaseFormValidator 
       .pipe(debounceTime(this._debounceTime))
       .pipe(takeUntil(this._unsubscribe))
       .subscribe(_ => (this._clusterSpecService.cluster = this._getClusterEntity()));
-
-    // Look-up for presets and floating IP pool is only required when we want to populate fields based on either the
-    // enabled preset or the provided credentials
-    if (!this.verifyCredentials) {
-      this.form.reset();
-      return;
-    }
-
-    this.form.get(Controls.FloatingIPPool).setValidators(Validators.required);
 
     this._presets.presetChanges.pipe(takeUntil(this._unsubscribe)).subscribe(preset =>
       Object.values(Controls).forEach(control => {
@@ -139,61 +87,12 @@ export class OpenstackApplicationCredentialsComponent extends BaseFormValidator 
         )
       );
 
-    merge(
-      this.form.get(Controls.ApplicationCredentialID).valueChanges.pipe(distinctUntilChanged()),
-      this.form.get(Controls.ApplicationCredentialSecret).valueChanges.pipe(distinctUntilChanged())
-    )
-      .pipe(debounceTime(this._debounceTime))
-      .pipe(tap(_ => this._clearFloatingIPPool()))
-      .pipe(switchMap(_ => this._floatingIPPoolListObservable()))
-      .pipe(takeUntil(this._unsubscribe))
-      .subscribe((floatingIPPools: OpenstackFloatingIPPool[]) => {
-        this.floatingIPPools = floatingIPPools;
-
-        if (!_.isEmpty(this.floatingIPPools)) {
-          this.floatingIPPoolsLabel = FloatingIPPoolState.Ready;
-          this._cdr.detectChanges();
-        }
-      });
-
-    merge(
-      this._clusterSpecService.providerChanges,
-      this._clusterSpecService.datacenterChanges,
-      this._credentialsTypeService.credentialsTypeChanges
-    )
-      .pipe(filter(_ => this._clusterSpecService.provider === NodeProvider.OPENSTACK))
-      .pipe(switchMap(_ => this._datacenterService.getDatacenter(this._clusterSpecService.datacenter).pipe(take(1))))
-      .pipe(tap(dc => (this._isFloatingPoolIPEnforced = dc.spec.openstack.enforceFloatingIP)))
-      .pipe(takeUntil(this._unsubscribe))
-      .subscribe(_ => this.form.reset());
     this.form.reset();
   }
 
   ngOnDestroy(): void {
     this._unsubscribe.next();
     this._unsubscribe.complete();
-  }
-
-  isRequired(control: Controls): boolean {
-    switch (control) {
-      case Controls.FloatingIPPool:
-        return this._isFloatingPoolIPEnforced;
-      default:
-        return true;
-    }
-  }
-
-  onFloatingIPPoolChange(floatingIPPool: string): void {
-    this._clusterSpecService.cluster.spec.cloud.openstack.floatingIPPool = floatingIPPool;
-  }
-
-  getHint(control: Controls): string {
-    switch (control) {
-      case Controls.FloatingIPPool:
-        return this._hasRequiredCredentials() ? '' : 'Please enter your credentials first.';
-    }
-
-    return '';
   }
 
   private _enable(enable: boolean, name: string): void {
@@ -204,48 +103,6 @@ export class OpenstackApplicationCredentialsComponent extends BaseFormValidator 
     if (!enable && this.form.get(name).enabled) {
       this.form.get(name).disable();
     }
-  }
-
-  private _hasRequiredCredentials(): boolean {
-    return (
-      !!this._clusterSpecService.cluster.spec.cloud.openstack &&
-      !!this._clusterSpecService.cluster.spec.cloud.openstack.applicationCredentialID &&
-      !!this._clusterSpecService.cluster.spec.cloud.openstack.applicationCredentialSecret
-    );
-  }
-
-  private _floatingIPPoolListObservable(): Observable<OpenstackFloatingIPPool[]> {
-    return this._presets
-      .provider(NodeProvider.OPENSTACK)
-      .applicationCredentialID(this.form.get(Controls.ApplicationCredentialID).value)
-      .applicationCredentialPassword(this.form.get(Controls.ApplicationCredentialSecret).value)
-      .datacenter(this._clusterSpecService.cluster.spec.cloud.dc)
-      .networks(this._onFloatingIPPoolLoading.bind(this))
-      .pipe(
-        map(networks => {
-          const filteredNetworks = networks.filter(network => network.external === true);
-          return _.sortBy(filteredNetworks, n => n.name.toLowerCase());
-        })
-      )
-      .pipe(
-        catchError(() => {
-          this._clearFloatingIPPool();
-          return onErrorResumeNext(EMPTY);
-        })
-      );
-  }
-
-  private _clearFloatingIPPool(): void {
-    this.floatingIPPools = [];
-    this._floatingIPPoolCombobox.reset();
-    this.floatingIPPoolsLabel = FloatingIPPoolState.Empty;
-    this._cdr.detectChanges();
-  }
-
-  private _onFloatingIPPoolLoading(): void {
-    this._clearFloatingIPPool();
-    this.floatingIPPoolsLabel = FloatingIPPoolState.Loading;
-    this._cdr.detectChanges();
   }
 
   private _getClusterEntity(): Cluster {
