+++ src/app/shared/components/openstack-credentials/template.html
@@ -1,5 +1,5 @@
 <!--
-Copyright 2022 The Kubermatic Kubernetes Platform contributors.
+Copyright 2020 The Kubermatic Kubernetes Platform contributors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -32,10 +32,11 @@ limitations under the License.
   </mat-button-toggle-group>
 
   <ng-container *ngIf="credentialsType === CredentialsType.Default">
-    <km-openstack-default-credentials [formControl]="form.get(Controls.Credentials)" [verifyCredentials]="verifyCredentials"></km-openstack-default-credentials>
+    <km-openstack-default-credentials [formControl]="form.get(Controls.Credentials)"
+                                      [mode]="mode"></km-openstack-default-credentials>
   </ng-container>
 
   <ng-container *ngIf="credentialsType === CredentialsType.Application">
-    <km-openstack-application-credentials [formControl]="form.get(Controls.Credentials)" [verifyCredentials]="verifyCredentials"></km-openstack-application-credentials>
+    <km-openstack-application-credentials [formControl]="form.get(Controls.Credentials)"></km-openstack-application-credentials>
   </ng-container>
 </form>
